requirements.txt
Flask
SQLAlchemy
pyjwt
bcrypt  